requirements.txt
numpy>=1.19.5
pandas>=1.3.0
torch>=1.8.0
torchvision>=0.9.0
transformers>=4.11.0
Pillow>=8.2.0
matplotlib>=3.4.2
seaborn>=0.11.1
scikit-learn>=0.24.2
ftfy>=6.0.3
regex>=2021.8.3
tqdm>=4.62.0
clip @ git+https://github.com/openai/CLIP.git
jupyterlab>=3.0.0
ipywidgets>=7.6.3 